NEW: move mm tests to different class
